split up and renamed pooling client
